backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.0
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.0
aioredis==2.0.1
numpy==1.24.0
torch==2.1.0
gymnasium==0.29.0
stable-baselines3==2.2.0
boto3==1.34.0
celery==5.3.0
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.12.0
mypy==1.7.0
httpx==0.25.0